archive/src_components/universal_llm_kit/requirements.txt
# Universal LLM Kit Requirements
litellm>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For enhanced structured output
instructor>=1.0.0

# Optional: For async support
aiohttp>=3.8.0

# For consensus system analysis and visualization
matplotlib>=3.5.0
numpy>=1.20.0